// src/app/models/admin.dashboard.ts
export interface AdminDashboard {
    total_users: number;

    total_artists: number;

    total_active_users: number;

    total_albums: number;

    total_approved_albums: number;

    total_pending_albums: number;

    total_songs: number;

    total_approved_songs: number;

    total_pending_songs: number;
}